+++ packages/native/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @regenrek/gpuix-native
 
+## 0.5.0-regenrek.11
+
 ## 0.5.0-regenrek.10
 
 ### Patch Changes